Routers done, killed bugs, answered questions, MVP
