+++ index.js
@@ -1,29 +1,25 @@
-const express = require('express');
-const helmet = require('helmet');
-// const cors = require('cors');
+/*
+play this: https://www.youtube.com/watch?v=d-diB65scQU
 
-const projectRouter = require('./data/routers/projectRouter');
-const actionRouter = require('../routers/actionRouter');
+Sing along:
 
-const server = express();
+here's a little code I wrote, you might want to read it really slow, don't worry be happy
+in every line there may be trouble, but if you worry you make it double, don't worry, be happy
+ain't got no sense of what is REST? just concentrate on learning Express, don't worry, be happy
+your file is getting way too big, bring a Router and make it thin, don't worry, be crafty
+there is no data on that route, just write some code, you'll sort it out… don't worry, just API…
+I need this code, just don't know where, perhaps should make some middleware, don't worry, be happy
 
-server.use(logger);
-server.use(helmet());
-server.use(express.json());
-// server.use(cors());
+Go code!
+*/
 
-server.use('/api/projects', projectRouter);
-server.use('/api/actions', actionRouter);
+require('dotenv').config();
+const server = require("./server");
 
-server.get('/', (req, res) => {
-  res.send(`<h2>Hello world from Express and Node.js Sprint Challenge!</h2>`);
-});
+// reading an environment variable
+const port = process.env.PORT || 4000;
+server.listen(port, () =>
+  console.log(`\n** Server running on http://localhost${port} **\n`),
+);
 
-// Custom Middleware
-function logger(req, res, next) {
-  console.log(`[${new Date().toISOString()}] ${req.method} to ${req.url}`);
 
-  next();
-}
-
-module.exports = server;
