+++ package.json
@@ -23,6 +23,7 @@
   "homepage": "https://github.com/LambdaSchool/webapi-challenge#readme",
   "dependencies": {
     "bcrypt": "^3.0.7",
+    "dotenv": "^8.2.0",
     "express": "^4.17.1",
     "helmet": "^3.21.2",
     "knex": "^0.20.3",
