+++ routers/projectRouter.js
@@ -1,8 +1,8 @@
 const express = require('express');
 
-const Projects = require('../helpers/projectModel');
+const Projects = require('../data/helpers/projectModel');
 // const Actions = require('../helpers/');
-
+const Actions = require('../data/helpers/actionModel');
 const router = express.Router();
